Fix 404 routing error when clicking on Members on 'Our Team'
