+++ package.json
@@ -49,6 +49,7 @@
 	"dependencies": {
 		"@types/dompurify": "^3.0.5",
 		"@types/marked": "^5.0.2",
+		"amplify-adapter": "^0.2.0",
 		"dompurify": "^3.1.7",
 		"lucide-svelte": "^0.453.0",
 		"marked": "^14.1.4",
