--- src/routes/Team/+page.ts
@@ -1,4 +1,0 @@
-// src/routes/Team/+page.ts
-export function load() {
-	return {};
-}
