+++ src/routes/Team/[member]/+page.server.ts
@@ -1,0 +1,5 @@
+export function load({ params }) {
+	return {
+		member: params.member
+	};
+}
