--- src/routes/Team/[slug]/+page.ts
@@ -1,6 +1,0 @@
-// src/routes/Team/[slug]/+page.ts
-export function load({ params }) {
-	return {
-		slug: params.slug
-	};
-}
