+++ svelte.config.js
@@ -1,5 +1,5 @@
-import adapter from '@sveltejs/adapter-auto';
 import { vitePreprocess } from '@sveltejs/vite-plugin-svelte';
+import adapter from 'amplify-adapter';
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
@@ -12,8 +12,7 @@ const config = {
 		// If your environment is not supported, or you settled on a specific environment, switch out the adapter.
 		// See https://kit.svelte.dev/docs/adapters for more information about adapters.
 		adapter: adapter()
-	},
-	preprocess: vitePreprocess()
+	}
 };
 
 export default config;
